requirements.txt
python-telegram-bot[job-queue]==21.*
aiosqlite
yt-dlp
httpx
python-dotenv
psutil
